requirements.txt
streamlit
boto3
PyPDF2
python-pptx
python-docx
langdetect
pillow
pandas
numpy
python-dotenv
streamlit-option-menu
plotly
matplotlib
seaborn

#awscli
